requirements.txt
absl-py==0.13.0
astunparse==1.6.3
attrs==21.2.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
flatbuffers==1.12
gast==0.4.0
google-api-core==1.30.0
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-cloud-speech==2.4.0
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
h5py==3.1.0
idna==2.9
imageio==2.9.0
imantics==0.1.12
imgaug==0.4.0
jsonschema==3.2.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
labelme2coco==0.1.2
libcst==0.3.19
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.4.2
mypy-extensions==0.4.3
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==20.9
Pillow==8.2.0
pixellib==0.6.1
proto-plus==1.18.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.23.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.18.1
scipy==1.6.3
Shapely==1.7.1
six==1.15.0
srt==3.0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tifffile==2021.6.14
typing-extensions==3.7.4.3
typing-inspect==0.7.1
urllib3==1.25.8
Werkzeug==2.0.1
wrapt==1.12.1
xmljson==0.2.1
